@@ -1,152 +1,151 @@
 using System;
 using System.Linq.Expressions;
 using System.Reflection;
-using Aop.Cache.ExpirationManagement;
+using Microsoft.Extensions.Caching.Memory;
 using Moq;
 using Xunit;
 
-namespace Aop.Cache.Unit.Tests
+namespace Aop.Cache.Unit.Tests;
+
+public class ExpectationTests
 {
-    public partial class ExpectationTests
+    [Fact]
+    public void SignatureMatchYieldsHit()
     {
-        [Fact]
-        public void SignatureMatchYieldsHit()
-        {
-            Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
-            var expiration = While.Result.True<object>((i, dt) => true);
+        Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
 
-            var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression, expiration);
+        var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression);
 
-            var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
-            var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
+        var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
+        var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
 
-            methodInfo.Setup(x => x.Name).Returns("MethodCall");
-            methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
+        methodInfo.Setup(x => x.Name).Returns("MethodCall");
+        methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
 
-            invocation.Setup(x => x.Method).Returns(methodInfo.Object);
-            invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
+        invocation.Setup(x => x.TargetType).Returns(typeof(ForTestingPurposes));
+        invocation.Setup(x => x.Method).Returns(methodInfo.Object);
+        invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
 
-            Assert.True(expectation.IsHit(invocation.Object));
-        }
+        Assert.True(expectation.IsHit(invocation.Object));
+    }
 
-        [Fact]
-        public void FuzzyMatchYieldsHit()
-        {
-            Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(It.IsAny<int>(), "zero");
-            var expiration = While.Result.True<object>((i, dt) => true);
+    [Fact]
+    public void FuzzyMatchYieldsHit()
+    {
+        Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(It.IsAny<int>(), "zero");
 
-            var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression, expiration);
+        var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression);
 
-            var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
-            var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
+        var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
+        var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
 
-            methodInfo.Setup(x => x.Name).Returns("MethodCall");
-            methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
+        methodInfo.Setup(x => x.Name).Returns("MethodCall");
+        methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
+        invocation.Setup(x => x.TargetType).Returns(typeof(ForTestingPurposes));
+        invocation.Setup(x => x.Method).Returns(methodInfo.Object);
+        invocation.Setup(x => x.Arguments).Returns(new object[] { 42, "zero" });
 
-            invocation.Setup(x => x.Method).Returns(methodInfo.Object);
-            invocation.Setup(x => x.Arguments).Returns(new object[] { 42, "zero" });
+        Assert.True(expectation.IsHit(invocation.Object));
+    }
 
-            Assert.True(expectation.IsHit(invocation.Object));
-        }
+    [Fact]
+    public void NotNullExpectationNotNullValueYieldsHit()
+    {
+        Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, It.IsNotNull<string>());
 
-        [Fact]
-        public void NotNullExpectationNotNullValueYieldsHit()
-        {
-            Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, It.IsNotNull<string>());
-            var expiration = While.Result.True<object>((i, dt) => true);
+        var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression);
 
-            var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression, expiration);
+        var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
+        var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
 
-            var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
-            var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
+        methodInfo.Setup(x => x.Name).Returns("MethodCall");
+        methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
 
-            methodInfo.Setup(x => x.Name).Returns("MethodCall");
-            methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
+        invocation.Setup(x => x.TargetType).Returns(typeof(ForTestingPurposes));
+        invocation.Setup(x => x.Method).Returns(methodInfo.Object);
+        invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
 
-            invocation.Setup(x => x.Method).Returns(methodInfo.Object);
-            invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
+        Assert.True(expectation.IsHit(invocation.Object));
+    }
 
-            Assert.True(expectation.IsHit(invocation.Object));
-        }
+    [Fact]
+    public void NotNullExpectationNullParameterYieldsNoHit()
+    {
+        Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, It.IsNotNull<string>());
 
-        [Fact]
-        public void NotNullExpectationNullParameterYieldsNoHit()
-        {
-            Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, It.IsNotNull<string>());
-            var expiration = While.Result.True<object>((i, dt) => true);
+        var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression);
 
-            var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression, expiration);
+        var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
+        var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
 
-            var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
-            var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
+        methodInfo.Setup(x => x.Name).Returns("MethodCall");
+        methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
 
-            methodInfo.Setup(x => x.Name).Returns("MethodCall");
-            methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
+        invocation.Setup(x => x.TargetType).Returns(typeof(ForTestingPurposes));
+        invocation.Setup(x => x.Method).Returns(methodInfo.Object);
+        invocation.Setup(x => x.Arguments).Returns(new object[] { 42, null });
 
-            invocation.Setup(x => x.Method).Returns(methodInfo.Object);
-            invocation.Setup(x => x.Arguments).Returns(new object[] { 42, null });
+        Assert.False(expectation.IsHit(invocation.Object));
+    }
 
-            Assert.False(expectation.IsHit(invocation.Object));
-        }
+    [Fact]
+    public void ReturnTypeMismatchYieldsNoHit()
+    {
+        Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
 
-        [Fact]
-        public void ReturnTypeMismatchYieldsNoHit()
-        {
-            Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
-            var expiration = While.Result.True<object>((i, dt) => true);
+        var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression);
 
-            var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression, expiration);
+        var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
+        var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
 
-            var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
-            var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
+        methodInfo.Setup(x => x.Name).Returns("MethodCall");
+        methodInfo.Setup(x => x.ReturnType).Returns(typeof(int));
 
-            methodInfo.Setup(x => x.Name).Returns("MethodCall");
-            methodInfo.Setup(x => x.ReturnType).Returns(typeof(int));
+        invocation.Setup(x => x.TargetType).Returns(typeof(ForTestingPurposes));
+        invocation.Setup(x => x.Method).Returns(methodInfo.Object);
+        invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
 
-            invocation.Setup(x => x.Method).Returns(methodInfo.Object);
-            invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
+        Assert.False(expectation.IsHit(invocation.Object));
+    }
 
-            Assert.False(expectation.IsHit(invocation.Object));
-        }
+    [Fact]
+    public void ParameterMismatchYieldsNoHit()
+    {
+        Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
 
-        [Fact]
-        public void ParameterMismatchYieldsNoHit()
-        {
-            Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
-            var expiration = While.Result.True<object>((i, dt) => true);
+        var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression);
 
-            var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression, expiration);
+        var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
+        var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
 
-            var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
-            var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
+        methodInfo.Setup(x => x.Name).Returns("MethodCall");
+        methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
 
-            methodInfo.Setup(x => x.Name).Returns("MethodCall");
-            methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
+        invocation.Setup(x => x.TargetType).Returns(typeof(ForTestingPurposes));
+        invocation.Setup(x => x.Method).Returns(methodInfo.Object);
+        invocation.Setup(x => x.Arguments).Returns(new object[] { 1, "zero" });
 
-            invocation.Setup(x => x.Method).Returns(methodInfo.Object);
-            invocation.Setup(x => x.Arguments).Returns(new object[] { 1, "zero" });
+        Assert.False(expectation.IsHit(invocation.Object));
+    }
 
-            Assert.False(expectation.IsHit(invocation.Object));
-        }
+    [Fact]
+    public void MethodNameMismatchYieldsNoHit()
+    {
+        Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
 
-        [Fact]
-        public void MethodNameMismatchYieldsNoHit()
-        {
-            Expression<Func<ForTestingPurposes, string>> expression = s => s.MethodCall(0, "zero");
-            var expiration = While.Result.True<object>((i, dt) => true);
+        var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression);
 
-            var expectation = Expectation.FromMethodCallExpression(expression.Body as MethodCallExpression, expiration);
+        var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
+        var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
 
-            var invocation = new Mock<Castle.DynamicProxy.IInvocation>(MockBehavior.Strict);
-            var methodInfo = new Mock<MethodInfo>(MockBehavior.Strict);
+        methodInfo.Setup(x => x.Name).Returns("DoOtherStuff");
+        methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
 
-            methodInfo.Setup(x => x.Name).Returns("DoOtherStuff");
-            methodInfo.Setup(x => x.ReturnType).Returns(typeof(string));
+        invocation.Setup(x => x.TargetType).Returns(typeof(ForTestingPurposes));
+        invocation.Setup(x => x.Method).Returns(methodInfo.Object);
+        invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
 
-            invocation.Setup(x => x.Method).Returns(methodInfo.Object);
-            invocation.Setup(x => x.Arguments).Returns(new object[] { 0, "zero" });
-
-            Assert.False(expectation.IsHit(invocation.Object));
-        }
+        Assert.False(expectation.IsHit(invocation.Object));
     }
 }
+
