@@ -1,15 +1,14 @@
-﻿namespace Aop.Cache
-{
-    public static class It
-    {
-        public static T IsAny<T>()
-        {
-            return default(T);
-        }
+﻿namespace Aop.Cache;
 
-        public static T IsNotNull<T>()
-        {
-            return default(T);
-        }
+public static class It
+{
+    public static T IsAny<T>()
+    {
+        return default;
+    }
+
+    public static T IsNotNull<T>()
+    {
+        return default;
     }
 }
