@@ -1,6 +1,5 @@
-﻿namespace Aop.Cache
+﻿namespace Aop.Cache;
+
+public interface IPerInstanceAdapter<T> : ICacheAdapter<T> where T : class
 {
-    public interface IPerInstanceAdapter<T> : ICacheAdapter<T> where T : class
-    {
-    }
 }
